requirements/local.txt
Django==2.0.2
django-cors-headers==2.2.0
django-extensions==2.0.0
django-redis==4.9.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
psycopg2-binary==2.7.4